First lab (3)
This code solves the problem of numbering cells on a strip of paper after its k folds.
The getnum(k, n) function recursively calculates the cell number on the strip before folding by its number n and the number of the folded strip k.
If k=1 (the strip is already rolled up to one cell), then n & 1 is returned - the remainder from dividing n by 2 (because the cell numbers start with 0).
Otherwise, if the number of the folded strip is odd (i.e. the strip is not folded in half), then the cell number on the first half of the strip (before folding) is calculated using the recursive call getnum(k-1, ((1 << k) - 1 - n) >> 1), where ((1 << k) - 1 - n) >> 1 - calculation of the cell number on the first half of the strip before folding. Then the number of the corresponding cell on the folded strip is calculated using the formula (1 << k) - 1 - the number of the first half + the number of the second half.
If the number of the folded strip is even, then the cell number on the half of the strip before folding is calculated using the recursive call getnum(k-1, n >> 1), and multiplied by 2.
The solution(k, index) function calculates the cell number on the collapsed strip by its index number and the number of the collapsed strip k.
If k=0 (there is no strip), then 0 is returned.
If k=1 (the strip is already collapsed to one cell), then index & 1 is returned - the remainder from dividing index by 2 (because the cell numbers start with 0).
Otherwise, if the cell number on the folded strip is odd, then the number of the corresponding cell on the half of the strip before folding is calculated, multiplied by 2 and 1 is added using the getnum function.
If the cell number on the folded strip is even, then the number of the corresponding cell on the half of the strip before folding is calculated, multiplied by 2 using the getnum function.
In the main program, the number of collapses k is entered and the numbering of all cells on the collapsed strip is output in the order they follow from left to right. To do this, it cycles through all possible cell numbers on the collapsed strip and for each number the solution function is called, which calculates the corresponding cell number on the collapsed strip.